requirements.txt
torch>=2.1.2
torchvision>=0.16.2
pillow~=10.4.0
setuptools~=65.5.1
numpy~=1.23.5
matplotlib~=3.9.2
scipy~=1.13.1
tqdm~=4.66.5
pandas~=2.2.3
jupyterlab
ipykernel
ipympl
